+++ R/na_values.R
@@ -8,7 +8,16 @@
 #' @details
 #' See [haven::labelled_spss()] for a presentation of SPSS's user defined missing values.
 #' Note that [base::is.na()] will return `TRUE` for user defined missing values.
-#' You can use [user_na_to_na()] to convert user defined missing values to `NA`.
+#'
+#' You can use [user_na_to_na()] to convert user defined missing values to regular `NA`.
+#' Note that any value label attached to a user defined missing value will be lost.
+#'
+#' The method [user_na_to_tagged_na()] will convert user defined missing values
+#' into [haven::tagged_na()], preserving value labels. Please note that [haven::tagged_na()]
+#' are defined only for double vectors. Therefore, integer `haven_labelled_spss`
+#' vectors will be converted into double `haven_labelled` vectors; and
+#' [user_na_to_tagged_na()] cannot be applied to a character `haven_labelled_spss`
+#' vector.
 #' @return
 #'   `na_values()` will return a vector of values that should also be considered as missing.
 #'   `na_range()` will return a numeric vector of length two giving the (inclusive)
@@ -239,7 +248,11 @@ set_na_range <- function(.data, ..., .values = NA, .strict = TRUE) {
   .data
 }
 
-
+#' @rdname na_values
+#' @export
+is_user_na <- function(x) {
+  test_if_user_na(x, na_values(x), na_range(x))
+}
 
 #' @rdname na_values
 #' @export
@@ -287,4 +300,16 @@ user_na_to_tagged_na.haven_labelled_spss <- function(x) {
 user_na_to_tagged_na.data.frame <- function(x) {
   x[] <- lapply(x, user_na_to_tagged_na)
   x
+}
+
+# internal function to test if a value is user_na
+test_if_user_na <- function(val, na_values = NULL, na_range = NULL) {
+  miss <- rep.int(FALSE, length(val))
+  if (!is.null(na_values)) {
+    miss <- miss | val %in% na_values
+  }
+  if (!is.null(na_range)) {
+    miss <- miss | (val >= na_range[1] & val <= na_range[2])
+  }
+  miss
 }